Transfer rule gaps tests to Detection Engine (API Integration)
