+++ x-pack/test/security_solution_api_integration/test_suites/detections_response/rules_management/rule_bulk_actions/trial_license_complete_tier/perform_bulk_action.ts
@@ -9,7 +9,6 @@ import expect from 'expect';
 import {
   DETECTION_ENGINE_RULES_BULK_ACTION,
   DETECTION_ENGINE_RULES_URL,
-  MAX_MANUAL_RULE_RUN_LOOKBACK_WINDOW_DAYS,
   NOTIFICATION_THROTTLE_RULE,
 } from '@kbn/security-solution-plugin/common/constants';
 import {
@@ -21,7 +20,6 @@ import { EXCEPTION_LIST_ITEM_URL, EXCEPTION_LIST_URL } from '@kbn/securitysoluti
 import { getCreateExceptionListItemMinimalSchemaMock } from '@kbn/lists-plugin/common/schemas/request/create_exception_list_item_schema.mock';
 import { AuthType } from '@kbn/stack-connectors-plugin/common/auth/constants';
 import { BaseDefaultableFields } from '@kbn/security-solution-plugin/common/api/detection_engine';
-import moment from 'moment';
 import {
   binaryToString,
   getSimpleMlRule,
@@ -38,9 +36,6 @@ import { createRule, deleteAllRules } from '../../../../../config/services/detec
 import { deleteAllExceptions } from '../../../../lists_and_exception_lists/utils';
 
 import { FtrProviderContext } from '../../../../../ftr_provider_context';
-import { deleteAllGaps } from '../../../utils/event_log/delete_all_gaps';
-import { GapEvent, generateGapsForRule } from '../../../utils/event_log/generate_gaps_for_rule';
-import { getGapsByRuleId } from '../../../../../config/services/detections_response/rules/get_gaps_by_rule_id';
 
 export default ({ getService }: FtrProviderContext): void => {
   const supertest = getService('supertest');
@@ -91,8 +86,7 @@ export default ({ getService }: FtrProviderContext): void => {
   const createWebHookConnector = () => createConnector(getWebHookAction());
   const createSlackConnector = () => createConnector(getSlackAction());
 
-  // Failing: See https://github.com/elastic/kibana/issues/224615
-  describe.skip('@ess @serverless @skipInServerless perform_bulk_action', () => {
+  describe('@ess @serverless @skipInServerless perform_bulk_action', () => {
     beforeEach(async () => {
       await deleteAllRules(supertest, log);
       await esArchiver.load('x-pack/platform/test/fixtures/es_archives/auditbeat/hosts');
@@ -2239,708 +2233,6 @@ export default ({ getService }: FtrProviderContext): void => {
       });
     });
 
-    // skipped on MKI since feature flags are not supported there
-    describe('@skipInServerlessMKI manual rule run action', () => {
-      it('should return all existing and enabled rules as succeeded', async () => {
-        const intervalInMinutes = 25;
-        const interval = `${intervalInMinutes}m`;
-        await createRule(
-          supertest,
-          log,
-          getCustomQueryRuleParams({
-            rule_id: 'rule-1',
-            enabled: true,
-            interval,
-          })
-        );
-        await createRule(
-          supertest,
-          log,
-          getCustomQueryRuleParams({
-            rule_id: 'rule-2',
-            enabled: true,
-            interval,
-          })
-        );
-
-        const endDate = moment();
-        const startDate = endDate.clone().subtract(1, 'h');
-
-        const { body } = await securitySolutionApi
-          .performRulesBulkAction({
-            query: {},
-            body: {
-              query: '',
-              action: BulkActionTypeEnum.run,
-              [BulkActionTypeEnum.run]: {
-                start_date: startDate.toISOString(),
-                end_date: endDate.toISOString(),
-              },
-            },
-          })
-          .expect(200);
-
-        expect(body.attributes.summary).toEqual({
-          failed: 0,
-          skipped: 0,
-          succeeded: 2,
-          total: 2,
-        });
-        expect(body.attributes.errors).toBeUndefined();
-      });
-
-      it('should return 400 error when start date > end date', async () => {
-        const intervalInMinutes = 25;
-        const interval = `${intervalInMinutes}m`;
-        const createdRule1 = await createRule(
-          supertest,
-          log,
-          getCustomQueryRuleParams({
-            rule_id: 'rule-1',
-            enabled: true,
-            interval,
-          })
-        );
-        const createdRule2 = await createRule(
-          supertest,
-          log,
-          getCustomQueryRuleParams({
-            rule_id: 'rule-2',
-            enabled: true,
-            interval,
-          })
-        );
-
-        const endDate = moment();
-        const startDate = endDate.clone().subtract(1, 'h');
-
-        const { body } = await securitySolutionApi
-          .performRulesBulkAction({
-            query: {},
-            body: {
-              ids: [createdRule1.id, createdRule2.id],
-              action: BulkActionTypeEnum.run,
-              [BulkActionTypeEnum.run]: {
-                start_date: endDate.toISOString(),
-                end_date: startDate.toISOString(),
-              },
-            },
-          })
-          .expect(400);
-
-        expect(body.message).toContain('[0]: Backfill end must be greater than backfill start');
-      });
-
-      it('should return 400 error when start date = end date', async () => {
-        const intervalInMinutes = 25;
-        const interval = `${intervalInMinutes}m`;
-        const createdRule1 = await createRule(
-          supertest,
-          log,
-          getCustomQueryRuleParams({
-            rule_id: 'rule-1',
-            enabled: true,
-            interval,
-          })
-        );
-        const createdRule2 = await createRule(
-          supertest,
-          log,
-          getCustomQueryRuleParams({
-            rule_id: 'rule-2',
-            enabled: true,
-            interval,
-          })
-        );
-
-        const endDate = moment().subtract(1, 'h');
-        const startDate = endDate.clone();
-
-        const { body } = await securitySolutionApi
-          .performRulesBulkAction({
-            query: {},
-            body: {
-              ids: [createdRule1.id, createdRule2.id],
-              action: BulkActionTypeEnum.run,
-              [BulkActionTypeEnum.run]: {
-                start_date: startDate.toISOString(),
-                end_date: endDate.toISOString(),
-              },
-            },
-          })
-          .expect(400);
-
-        expect(body.message).toContain('[0]: Backfill end must be greater than backfill start');
-      });
-
-      it('should return 400 error when start date is in the future', async () => {
-        const intervalInMinutes = 25;
-        const interval = `${intervalInMinutes}m`;
-        const createdRule1 = await createRule(
-          supertest,
-          log,
-          getCustomQueryRuleParams({
-            rule_id: 'rule-1',
-            enabled: true,
-            interval,
-          })
-        );
-        const createdRule2 = await createRule(
-          supertest,
-          log,
-          getCustomQueryRuleParams({
-            rule_id: 'rule-2',
-            enabled: true,
-            interval,
-          })
-        );
-
-        const startDate = moment().add(1, 'd');
-        const endDate = moment().add(2, 'd');
-
-        const { body } = await securitySolutionApi
-          .performRulesBulkAction({
-            query: {},
-            body: {
-              ids: [createdRule1.id, createdRule2.id],
-              action: BulkActionTypeEnum.run,
-              [BulkActionTypeEnum.run]: {
-                start_date: startDate.toISOString(),
-                end_date: endDate.toISOString(),
-              },
-            },
-          })
-          .expect(400);
-
-        expect(body.message).toContain('[0]: Backfill cannot be scheduled for the future');
-      });
-
-      it('should return 400 error when end date is in the future', async () => {
-        const intervalInMinutes = 25;
-        const interval = `${intervalInMinutes}m`;
-        const createdRule1 = await createRule(
-          supertest,
-          log,
-          getCustomQueryRuleParams({
-            rule_id: 'rule-1',
-            enabled: true,
-            interval,
-          })
-        );
-        const createdRule2 = await createRule(
-          supertest,
-          log,
-          getCustomQueryRuleParams({
-            rule_id: 'rule-2',
-            enabled: true,
-            interval,
-          })
-        );
-
-        const endDate = moment().add(1, 'd');
-        const startDate = moment().subtract(1, 'd');
-
-        const { body } = await securitySolutionApi
-          .performRulesBulkAction({
-            query: {},
-            body: {
-              ids: [createdRule1.id, createdRule2.id],
-              action: BulkActionTypeEnum.run,
-              [BulkActionTypeEnum.run]: {
-                start_date: startDate.toISOString(),
-                end_date: endDate.toISOString(),
-              },
-            },
-          })
-          .expect(400);
-
-        expect(body.message).toContain('[0]: Backfill cannot be scheduled for the future');
-      });
-
-      it('should return 400 error when start date is far in the past', async () => {
-        const intervalInMinutes = 25;
-        const interval = `${intervalInMinutes}m`;
-        const createdRule1 = await createRule(
-          supertest,
-          log,
-          getCustomQueryRuleParams({
-            rule_id: 'rule-1',
-            enabled: true,
-            interval,
-          })
-        );
-        const createdRule2 = await createRule(
-          supertest,
-          log,
-          getCustomQueryRuleParams({
-            rule_id: 'rule-2',
-            enabled: true,
-            interval,
-          })
-        );
-
-        const endDate = moment();
-        const startDate = moment().subtract(MAX_MANUAL_RULE_RUN_LOOKBACK_WINDOW_DAYS + 1, 'd');
-
-        const { body } = await securitySolutionApi
-          .performRulesBulkAction({
-            query: {},
-            body: {
-              ids: [createdRule1.id, createdRule2.id],
-              action: BulkActionTypeEnum.run,
-              [BulkActionTypeEnum.run]: {
-                start_date: startDate.toISOString(),
-                end_date: endDate.toISOString(),
-              },
-            },
-          })
-          .expect(400);
-
-        expect(body.message).toContain(
-          `[0]: Backfill cannot look back more than ${MAX_MANUAL_RULE_RUN_LOOKBACK_WINDOW_DAYS} days`
-        );
-      });
-
-      it('should return 500 error if some rules do not exist', async () => {
-        const intervalInMinutes = 25;
-        const interval = `${intervalInMinutes}m`;
-        const createdRule1 = await createRule(
-          supertest,
-          log,
-          getCustomQueryRuleParams({
-            rule_id: 'rule-1',
-            enabled: true,
-            interval,
-          })
-        );
-
-        const endDate = moment();
-        const startDate = endDate.clone().subtract(1, 'h');
-
-        const { body } = await securitySolutionApi
-          .performRulesBulkAction({
-            query: {},
-            body: {
-              ids: [createdRule1.id, 'rule-2'],
-              action: BulkActionTypeEnum.run,
-              [BulkActionTypeEnum.run]: {
-                start_date: startDate.toISOString(),
-                end_date: endDate.toISOString(),
-              },
-            },
-          })
-          .expect(500);
-
-        expect(body.attributes.summary).toEqual({
-          failed: 1,
-          skipped: 0,
-          succeeded: 1,
-          total: 2,
-        });
-
-        expect(body.attributes.errors).toHaveLength(1);
-        expect(body.attributes.errors[0]).toEqual({
-          message: 'Rule not found',
-          status_code: 500,
-          rules: [
-            {
-              id: 'rule-2',
-            },
-          ],
-        });
-      });
-
-      it('should return 500 error if some rules are disabled', async () => {
-        const intervalInMinutes = 25;
-        const interval = `${intervalInMinutes}m`;
-        const createdRule1 = await createRule(
-          supertest,
-          log,
-          getCustomQueryRuleParams({
-            rule_id: 'rule-1',
-            enabled: false,
-            interval,
-          })
-        );
-        const createdRule2 = await createRule(
-          supertest,
-          log,
-          getCustomQueryRuleParams({
-            rule_id: 'rule-2',
-            enabled: true,
-            interval,
-          })
-        );
-
-        const endDate = moment();
-        const startDate = endDate.clone().subtract(1, 'h');
-
-        const { body } = await securitySolutionApi
-          .performRulesBulkAction({
-            query: {},
-            body: {
-              ids: [createdRule1.id, createdRule2.id],
-              action: BulkActionTypeEnum.run,
-              [BulkActionTypeEnum.run]: {
-                start_date: startDate.toISOString(),
-                end_date: endDate.toISOString(),
-              },
-            },
-          })
-          .expect(500);
-
-        expect(body.attributes.summary).toEqual({
-          failed: 1,
-          skipped: 0,
-          succeeded: 1,
-          total: 2,
-        });
-
-        expect(body.attributes.errors).toHaveLength(1);
-        expect(body.attributes.errors).toEqual(
-          expect.arrayContaining([
-            {
-              message: 'Cannot schedule manual rule run for a disabled rule',
-              status_code: 500,
-              err_code: 'MANUAL_RULE_RUN_DISABLED_RULE',
-              rules: [{ id: createdRule1.id, name: createdRule1.name }],
-            },
-          ])
-        );
-        expect(body.attributes.results).toEqual({
-          updated: [expect.objectContaining(createdRule2)],
-          created: [],
-          deleted: [],
-          skipped: [],
-        });
-      });
-    });
-
-    describe('@skipInServerlessMKI fill gaps run action', () => {
-      const intervalInMinutes = 5;
-      const interval = `${intervalInMinutes}m`;
-      const totalRules = 3;
-      type GapFromEvent = GapEvent['kibana']['alert']['rule']['gap'] & { _id: string };
-      let generatedGapEvents: Record<
-        string,
-        { rule: { id: string; name: string }; gapEvents: GapFromEvent[] }
-      >;
-
-      let backfillStart: Date;
-      let backfillEnd: Date;
-      let createdRules: Array<Awaited<ReturnType<typeof createRule>>>;
-      let createdRuleIds: string[];
-
-      const resetEverything = async () => {
-        await deleteAllGaps(es);
-        await deleteAllRules(supertest, log);
-      };
-
-      afterEach(resetEverything);
-
-      beforeEach(async () => {
-        createdRules = [];
-
-        for (let idx = 0; idx < totalRules; idx++) {
-          const rule = await createRule(
-            supertest,
-            log,
-            getCustomQueryRuleParams({
-              rule_id: idx.toString(),
-              enabled: true,
-              interval,
-            }),
-            'default'
-          );
-
-          createdRules.push(rule);
-
-          backfillEnd = new Date();
-          backfillStart = new Date(backfillEnd.getTime() - 24 * 60 * 60 * 1000);
-        }
-
-        createdRuleIds = createdRules.map(({ id }) => id);
-      });
-
-      describe('scheduling gap fills for rules', () => {
-        beforeEach(async () => {
-          generatedGapEvents = {};
-          for (const rule of createdRules) {
-            const { gapEvents } = await generateGapsForRule(es, rule, 100);
-            generatedGapEvents[rule.id] = {
-              rule,
-              gapEvents: gapEvents.map((gapEvent) => {
-                if (!gapEvent._id) {
-                  throw new Error('generated gap event id cannot be undefined');
-                }
-                return { ...gapEvent.kibana.alert.rule.gap, _id: gapEvent._id };
-              }),
-            };
-          }
-
-          let earliest = Date.now();
-          let latest = 0;
-
-          Object.values(generatedGapEvents).forEach(({ gapEvents }) => {
-            gapEvents
-              .flatMap((event) => event.unfilled_intervals)
-              .forEach(({ gte, lte }) => {
-                earliest = Math.min(earliest, new Date(gte).getTime());
-                latest = Math.max(latest, new Date(lte).getTime());
-              });
-          });
-
-          backfillEnd = new Date(latest);
-          backfillStart = new Date(earliest);
-        });
-
-        it('should trigger the scheduling of gap fills for the rules in the request', async () => {
-          // Only backfill the first 2 rules
-          const ruleIdsToBackfill = Object.keys(generatedGapEvents).slice(0, 2);
-
-          // Trigger the backfill for the selected rules
-          const { body } = await securitySolutionApi
-            .performRulesBulkAction({
-              query: {},
-              body: {
-                ids: ruleIdsToBackfill,
-                action: BulkActionTypeEnum.fill_gaps,
-                [BulkActionTypeEnum.fill_gaps]: {
-                  start_date: backfillStart.toISOString(),
-                  end_date: backfillEnd.toISOString(),
-                },
-              },
-            })
-            .expect(200);
-
-          expect(body.success).toEqual(true);
-          expect(body.attributes.summary).toEqual({
-            failed: 0,
-            succeeded: 2,
-            skipped: 0,
-            total: 2,
-          });
-
-          const expectedUpdatedRules = Object.values(generatedGapEvents)
-            .slice(0, 2)
-            .map((event) => event.rule);
-
-          expect(body.attributes.results).toEqual({
-            updated: expect.arrayContaining(
-              expectedUpdatedRules.map((expected) => expect.objectContaining(expected))
-            ),
-            created: [],
-            deleted: [],
-            skipped: [],
-          });
-
-          for (const ruleId of ruleIdsToBackfill) {
-            const fetchedGaps = await getGapsByRuleId(
-              supertest,
-              ruleId,
-              { start: backfillStart.toISOString(), end: backfillEnd.toISOString() },
-              100
-            );
-
-            const generatedGaps = generatedGapEvents[ruleId].gapEvents;
-
-            // Verify that every single gap is marked as in progress
-            generatedGaps.forEach((generatedGap) => {
-              const fetchedGap = fetchedGaps.find(({ _id }) => _id === generatedGap._id);
-              expect(fetchedGap?.unfilled_intervals).toEqual([]);
-              expect(fetchedGap?.in_progress_intervals).toEqual(generatedGap.unfilled_intervals);
-            });
-          }
-
-          // For the rules we didn't backfill, verify that their gaps are still unfilled
-          for (const ruleId of Object.keys(generatedGapEvents).slice(2)) {
-            const fetchedGaps = await getGapsByRuleId(
-              supertest,
-              ruleId,
-              { start: backfillStart.toISOString(), end: backfillEnd.toISOString() },
-              100
-            );
-
-            const generatedGaps = generatedGapEvents[ruleId].gapEvents;
-
-            generatedGaps.forEach((generatedGap) => {
-              const fetchedGap = fetchedGaps.find(({ _id }) => _id === generatedGap._id);
-              expect(fetchedGap?.unfilled_intervals).toEqual(generatedGap.unfilled_intervals);
-              expect(fetchedGap?.in_progress_intervals).toEqual([]);
-            });
-          }
-        });
-
-        it('should return 500 error if some rules do not exist', async () => {
-          const existentRules = createdRuleIds;
-          const nonExistentRule = 'non-existent-rule';
-          const { body } = await securitySolutionApi
-            .performRulesBulkAction({
-              query: {},
-              body: {
-                ids: [...existentRules, nonExistentRule],
-                action: BulkActionTypeEnum.fill_gaps,
-                [BulkActionTypeEnum.fill_gaps]: {
-                  start_date: backfillStart.toISOString(),
-                  end_date: backfillEnd.toISOString(),
-                },
-              },
-            })
-            .expect(500);
-
-          expect(body.attributes.summary).toEqual({
-            failed: 1,
-            skipped: 0,
-            succeeded: existentRules.length,
-            total: existentRules.length + 1,
-          });
-
-          expect(body.attributes.errors).toHaveLength(1);
-          expect(body.attributes.errors[0]).toEqual({
-            message: 'Rule not found',
-            status_code: 500,
-            rules: [
-              {
-                id: nonExistentRule,
-              },
-            ],
-          });
-        });
-
-        it('should return 500 error if some rules are disabled', async () => {
-          const enabledRules = createdRuleIds;
-          const disabledRule = await createRule(
-            supertest,
-            log,
-            getCustomQueryRuleParams({
-              rule_id: 'rule-disabled',
-              enabled: false,
-              interval,
-            })
-          );
-
-          await generateGapsForRule(es, disabledRule, 100);
-
-          const { body } = await securitySolutionApi
-            .performRulesBulkAction({
-              query: {},
-              body: {
-                ids: [...enabledRules, disabledRule.id],
-                action: BulkActionTypeEnum.fill_gaps,
-                [BulkActionTypeEnum.fill_gaps]: {
-                  start_date: backfillStart.toISOString(),
-                  end_date: backfillEnd.toISOString(),
-                },
-              },
-            })
-            .expect(500);
-
-          expect(body.attributes.summary).toEqual({
-            failed: 1,
-            skipped: 0,
-            succeeded: enabledRules.length,
-            total: enabledRules.length + 1,
-          });
-
-          expect(body.attributes.errors).toHaveLength(1);
-          expect(body.attributes.errors).toEqual(
-            expect.arrayContaining([
-              {
-                message: 'Cannot bulk fill gaps for a disabled rule',
-                status_code: 500,
-                err_code: 'RULE_FILL_GAPS_DISABLED_RULE',
-                rules: [{ id: disabledRule.id, name: disabledRule.name }],
-              },
-            ])
-          );
-
-          const expectedUpdatedRules = Object.values(generatedGapEvents).map((event) => event.rule);
-          expect(body.attributes.results).toEqual({
-            updated: expect.arrayContaining(
-              expectedUpdatedRules.map((expected) => expect.objectContaining(expected))
-            ),
-            created: [],
-            deleted: [],
-            skipped: [],
-          });
-        });
-      });
-
-      it('should return 400 error when the end date is not strictly greater than the start date', async () => {
-        const { body } = await securitySolutionApi
-          .performRulesBulkAction({
-            query: {},
-            body: {
-              ids: createdRuleIds,
-              action: BulkActionTypeEnum.fill_gaps,
-              [BulkActionTypeEnum.fill_gaps]: {
-                start_date: backfillStart.toISOString(),
-                end_date: backfillStart.toISOString(),
-              },
-            },
-          })
-          .expect(400);
-
-        expect(body.message).toContain('Backfill end must be greater than backfill start');
-      });
-
-      it('should return 400 error when start date is in the future', async () => {
-        const { body } = await securitySolutionApi
-          .performRulesBulkAction({
-            query: {},
-            body: {
-              ids: createdRuleIds,
-              action: BulkActionTypeEnum.fill_gaps,
-              [BulkActionTypeEnum.fill_gaps]: {
-                start_date: new Date(Date.now() + 1000).toISOString(),
-                end_date: backfillEnd.toISOString(),
-              },
-            },
-          })
-          .expect(400);
-
-        expect(body.message).toContain('Backfill cannot be scheduled for the future');
-      });
-
-      it('should return 400 error when end date is in the future', async () => {
-        const { body } = await securitySolutionApi
-          .performRulesBulkAction({
-            query: {},
-            body: {
-              ids: createdRuleIds,
-              action: BulkActionTypeEnum.fill_gaps,
-              [BulkActionTypeEnum.fill_gaps]: {
-                start_date: backfillStart.toISOString(),
-                end_date: new Date(Date.now() + 5 * 60 * 1000).toISOString(),
-              },
-            },
-          })
-          .expect(400);
-
-        expect(body.message).toContain('Backfill cannot be scheduled for the future');
-      });
-
-      it('should return 400 error when range between start and end are greater than 90 days', async () => {
-        const { body } = await securitySolutionApi
-          .performRulesBulkAction({
-            query: {},
-            body: {
-              ids: createdRuleIds,
-              action: BulkActionTypeEnum.fill_gaps,
-              [BulkActionTypeEnum.fill_gaps]: {
-                start_date: new Date(
-                  backfillEnd.getTime() - 1 - 90 * 24 * 60 * 60 * 1000
-                ).toISOString(),
-                end_date: backfillEnd.toISOString(),
-              },
-            },
-          })
-          .expect(400);
-
-        expect(body.message).toContain('Backfill cannot look back more than 90 days');
-      });
-    });
-
     describe('overwrite_data_views', () => {
       it('should add an index pattern to a rule and overwrite the data view when overwrite_data_views is true', async () => {
         const ruleId = 'ruleId';
@@ -3380,35 +2672,6 @@ export default ({ getService }: FtrProviderContext): void => {
 
         expect(skippedRule.index).toEqual(['index1-*']);
         expect(skippedRule.tags).toEqual(['tag1', 'tag2']);
-      });
-    });
-
-    describe('gaps_range filtering', () => {
-      it('should not affect rules without gaps when using gaps_range filters', async () => {
-        // Create two rules without gaps
-        await createRule(supertest, log, {
-          ...getSimpleRule('rule-without-gaps-1'),
-        });
-        await createRule(supertest, log, {
-          ...getSimpleRule('rule-without-gaps-2'),
-        });
-
-        // Execute bulk action with gaps range filter
-        const { body } = await postBulkAction().send({
-          query: '',
-          action: BulkActionTypeEnum.duplicate,
-          gaps_range_start: '2025-01-01T00:00:00.000Z',
-          gaps_range_end: '2025-01-02T00:00:00.000Z',
-          duplicate: { include_exceptions: false, include_expired_exceptions: false },
-        });
-
-        // Verify the summary shows no rules were processed
-        expect(body.attributes.summary).toEqual({
-          failed: 0,
-          skipped: 0,
-          succeeded: 0,
-          total: 0,
-        });
       });
     });
