+++ x-pack/test/security_solution_api_integration/test_suites/detections_response/rules_management/rule_bulk_actions/trial_license_complete_tier/perform_bulk_action_dry_run.ts
@@ -9,7 +9,6 @@ import {
   BulkActionTypeEnum,
   BulkActionEditTypeEnum,
 } from '@kbn/security-solution-plugin/common/api/detection_engine/rule_management';
-import moment from 'moment';
 import {
   getCustomQueryRuleParams,
   getSimpleMlRule,
@@ -249,324 +248,6 @@ export default ({ getService }: FtrProviderContext): void => {
               ],
             });
           });
-        });
-      });
-    });
-
-    describe('@skipInServerless @skipInServerlessMKI schedule manual rule run action', () => {
-      it('should return all existing and enabled rules as succeeded', async () => {
-        const intervalInMinutes = 25;
-        const interval = `${intervalInMinutes}m`;
-        const createdRule1 = await createRule(
-          supertest,
-          log,
-          getCustomQueryRuleParams({
-            rule_id: 'rule-1',
-            enabled: true,
-            interval,
-          })
-        );
-        const createdRule2 = await createRule(
-          supertest,
-          log,
-          getCustomQueryRuleParams({
-            rule_id: 'rule-2',
-            enabled: true,
-            interval,
-          })
-        );
-
-        const endDate = moment();
-        const startDate = endDate.clone().subtract(1, 'h');
-
-        const { body } = await securitySolutionApi
-          .performRulesBulkAction({
-            query: { dry_run: true },
-            body: {
-              ids: [createdRule1.id, createdRule2.id],
-              action: BulkActionTypeEnum.run,
-              [BulkActionTypeEnum.run]: {
-                start_date: startDate.toISOString(),
-                end_date: endDate.toISOString(),
-              },
-            },
-          })
-          .expect(200);
-
-        expect(body.attributes.summary).toEqual({
-          failed: 0,
-          skipped: 0,
-          succeeded: 2,
-          total: 2,
-        });
-        expect(body.attributes.errors).toBeUndefined();
-      });
-
-      it('should return 500 error if some rules do not exist', async () => {
-        const intervalInMinutes = 25;
-        const interval = `${intervalInMinutes}m`;
-        const createdRule1 = await createRule(
-          supertest,
-          log,
-          getCustomQueryRuleParams({
-            rule_id: 'rule-1',
-            enabled: true,
-            interval,
-          })
-        );
-
-        const endDate = moment();
-        const startDate = endDate.clone().subtract(1, 'h');
-
-        const { body } = await securitySolutionApi
-          .performRulesBulkAction({
-            query: { dry_run: true },
-            body: {
-              ids: [createdRule1.id, 'rule-2'],
-              action: BulkActionTypeEnum.run,
-              [BulkActionTypeEnum.run]: {
-                start_date: startDate.toISOString(),
-                end_date: endDate.toISOString(),
-              },
-            },
-          })
-          .expect(500);
-
-        expect(body.attributes.summary).toEqual({
-          failed: 1,
-          skipped: 0,
-          succeeded: 1,
-          total: 2,
-        });
-
-        expect(body.attributes.errors).toHaveLength(1);
-        expect(body.attributes.errors[0]).toEqual({
-          message: 'Rule not found',
-          status_code: 500,
-          rules: [
-            {
-              id: 'rule-2',
-            },
-          ],
-        });
-      });
-
-      it('should return 500 error if some rules are disabled', async () => {
-        const intervalInMinutes = 25;
-        const interval = `${intervalInMinutes}m`;
-        const createdRule1 = await createRule(
-          supertest,
-          log,
-          getCustomQueryRuleParams({
-            rule_id: 'rule-1',
-            enabled: false,
-            interval,
-          })
-        );
-        const createdRule2 = await createRule(
-          supertest,
-          log,
-          getCustomQueryRuleParams({
-            rule_id: 'rule-2',
-            enabled: true,
-            interval,
-          })
-        );
-
-        const endDate = moment();
-        const startDate = endDate.clone().subtract(1, 'h');
-
-        const { body } = await securitySolutionApi
-          .performRulesBulkAction({
-            query: { dry_run: true },
-            body: {
-              ids: [createdRule1.id, createdRule2.id],
-              action: BulkActionTypeEnum.run,
-              [BulkActionTypeEnum.run]: {
-                start_date: startDate.toISOString(),
-                end_date: endDate.toISOString(),
-              },
-            },
-          })
-          .expect(500);
-
-        expect(body.attributes.summary).toEqual({
-          failed: 1,
-          skipped: 0,
-          succeeded: 1,
-          total: 2,
-        });
-
-        expect(body.attributes.errors).toHaveLength(1);
-        expect(body.attributes.errors[0]).toEqual({
-          err_code: 'MANUAL_RULE_RUN_DISABLED_RULE',
-          message: 'Cannot schedule manual rule run for a disabled rule',
-          status_code: 500,
-          rules: [
-            {
-              id: createdRule1.id,
-              name: createdRule1.name,
-            },
-          ],
-        });
-      });
-    });
-
-    describe('@skipInServerless @skipInServerlessMKI schedule bulk gap filling action', () => {
-      it('should return all existing and enabled rules as succeeded', async () => {
-        const intervalInMinutes = 25;
-        const interval = `${intervalInMinutes}m`;
-        const createdRule1 = await createRule(
-          supertest,
-          log,
-          getCustomQueryRuleParams({
-            rule_id: 'rule-1',
-            enabled: true,
-            interval,
-          })
-        );
-        const createdRule2 = await createRule(
-          supertest,
-          log,
-          getCustomQueryRuleParams({
-            rule_id: 'rule-2',
-            enabled: true,
-            interval,
-          })
-        );
-
-        const endDate = moment();
-        const startDate = endDate.clone().subtract(1, 'h');
-
-        const { body } = await securitySolutionApi
-          .performRulesBulkAction({
-            query: { dry_run: true },
-            body: {
-              ids: [createdRule1.id, createdRule2.id],
-              action: BulkActionTypeEnum.fill_gaps,
-              [BulkActionTypeEnum.fill_gaps]: {
-                start_date: startDate.toISOString(),
-                end_date: endDate.toISOString(),
-              },
-            },
-          })
-          .expect(200);
-
-        expect(body.attributes.summary).toEqual({
-          failed: 0,
-          skipped: 0,
-          succeeded: 2,
-          total: 2,
-        });
-        expect(body.attributes.errors).toBeUndefined();
-      });
-
-      it('should return 500 error if some rules do not exist', async () => {
-        const intervalInMinutes = 25;
-        const interval = `${intervalInMinutes}m`;
-        const createdRule1 = await createRule(
-          supertest,
-          log,
-          getCustomQueryRuleParams({
-            rule_id: 'rule-1',
-            enabled: true,
-            interval,
-          })
-        );
-
-        const endDate = moment();
-        const startDate = endDate.clone().subtract(1, 'h');
-
-        const { body } = await securitySolutionApi
-          .performRulesBulkAction({
-            query: { dry_run: true },
-            body: {
-              ids: [createdRule1.id, 'rule-2'],
-              action: BulkActionTypeEnum.fill_gaps,
-              [BulkActionTypeEnum.fill_gaps]: {
-                start_date: startDate.toISOString(),
-                end_date: endDate.toISOString(),
-              },
-            },
-          })
-          .expect(500);
-
-        expect(body.attributes.summary).toEqual({
-          failed: 1,
-          skipped: 0,
-          succeeded: 1,
-          total: 2,
-        });
-
-        expect(body.attributes.errors).toHaveLength(1);
-        expect(body.attributes.errors[0]).toEqual({
-          message: 'Rule not found',
-          status_code: 500,
-          rules: [
-            {
-              id: 'rule-2',
-            },
-          ],
-        });
-      });
-
-      it('should return 500 error if some rules are disabled', async () => {
-        const intervalInMinutes = 25;
-        const interval = `${intervalInMinutes}m`;
-        const createdRule1 = await createRule(
-          supertest,
-          log,
-          getCustomQueryRuleParams({
-            rule_id: 'rule-1',
-            enabled: false,
-            interval,
-          })
-        );
-        const createdRule2 = await createRule(
-          supertest,
-          log,
-          getCustomQueryRuleParams({
-            rule_id: 'rule-2',
-            enabled: true,
-            interval,
-          })
-        );
-
-        const endDate = moment();
-        const startDate = endDate.clone().subtract(1, 'h');
-
-        const { body } = await securitySolutionApi
-          .performRulesBulkAction({
-            query: { dry_run: true },
-            body: {
-              ids: [createdRule1.id, createdRule2.id],
-              action: BulkActionTypeEnum.fill_gaps,
-              [BulkActionTypeEnum.fill_gaps]: {
-                start_date: startDate.toISOString(),
-                end_date: endDate.toISOString(),
-              },
-            },
-          })
-          .expect(500);
-
-        expect(body.attributes.summary).toEqual({
-          failed: 1,
-          skipped: 0,
-          succeeded: 1,
-          total: 2,
-        });
-
-        expect(body.attributes.errors).toHaveLength(1);
-        expect(body.attributes.errors[0]).toEqual({
-          err_code: 'RULE_FILL_GAPS_DISABLED_RULE',
-          message: 'Cannot bulk fill gaps for a disabled rule',
-          status_code: 500,
-          rules: [
-            {
-              id: createdRule1.id,
-              name: createdRule1.name,
-            },
-          ],
         });
       });
     });
